+++ src/main/java/com/br/requirementhub/config/JwtAuthenticationFilter.java
@@ -2,6 +2,7 @@ package com.br.requirementhub.config;
 
 import com.br.requirementhub.model.User;
 import com.br.requirementhub.repository.UserRepository;
+import com.br.requirementhub.services.JwtService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
